:sparkles: Add `HexColor` extension to convert `Color` to hexadecimal value
The commit adds a new file `util.dart` which contains an extension `HexColor` on `Color` class. This extension provides a method `toHex` to convert a `Color` object to its hexadecimal representation. It also includes an internal method `_generateAlpha` to generate the alpha value if needed.

The `toHex` method accepts two optional parameters: `leadingHashSign` and `withAlpha`. The `leadingHashSign` parameter determines whether to include the hash sign '#' in the hexadecimal value. The `withAlpha` parameter determines whether to include the alpha value in the hexadecimal value.

This extension will be used to convert colors to hexadecimal values in other parts of the codebase.

Affected files:
- `lib/src/tapioca_ball.dart`
- `lib/src/util.dart`